@@ -4,13 +4,16 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: any
+  geolocator: ^9.0.2
+  http: ^1.1.0
+  flutter_spinkit: ^5.2.0
 
 dev_dependencies:
   flutter_test:
