@@ -1,4 +1,7 @@
+import 'package:clima/services/weather.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:clima/screens/location_screen.dart';
 
 class LoadingScreen extends StatefulWidget {
   @override
@@ -7,15 +10,31 @@ class LoadingScreen extends StatefulWidget {
 
 class _LoadingScreenState extends State<LoadingScreen> {
   @override
+  initState(){
+    super.initState();
+    getLocationData();
+  }
+
+  void getLocationData() async {
+    // WeatherModel weatherModel = WeatherModel();
+    // weatherModel.getLocationWeather();
+    var weatherData = await WeatherModel().getLocationWeather(); // same as above but initialized inline
+
+    Navigator.push(context,
+        MaterialPageRoute(builder: (context) {
+          return LocationScreen(locationWeather: weatherData);
+        }
+    ));
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
-        child: RaisedButton(
-          onPressed: () {
-            //Get the current location
-          },
-          child: Text('Get Location'),
-        ),
+          child: SpinKitDoubleBounce(
+            color: Colors.white,
+            size: 100.0,
+          )
       ),
     );
   }
